fix: static cast

+++ client/src/Utils.cpp
@@ -16,7 +16,7 @@ AnimationConfig actionToAnimationConfig(CurrentAction action, Coordinate entityL
     case ActionType::Attack:
         return AnimationConfig{targetDirection, AnimationType::ATTACK, static_cast<float>(action.durationMs), action.looping};
     case ActionType::Respawn:
-        return AnimationConfig{{0, 0, 0}, AnimationType::DEATH, action.durationMs, action.looping};
+        return AnimationConfig{{0, 0, 0}, AnimationType::DEATH, static_cast<float>(action.durationMs), action.looping};
     default:
         return AnimationConfig{{0, 0, 0}, AnimationType::IDLE, 1000.f, true};
     }
